Change group and system assets
