+++ Packages/UGF.Module.Update/Runtime/UpdateGroupAsset.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 2ada8f43f7184bc99ae625dd1b6df012
-timeCreated: 1606769504
+guid: 41606afaaecc41cfa6bbad84400f3928
+timeCreated: 1606768102
