--- Packages/UGF.Module.Update/Runtime/UpdateGroupAssetBase.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 41606afaaecc41cfa6bbad84400f3928
-timeCreated: 1606768102
