+++ Packages/UGF.Module.Update/Runtime/UpdateSystemDescriptionAsset.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: ea71b94d072e454e90f7c508a0972ae1
-timeCreated: 1605643696
+guid: 10d61bf07f3d4bbea735055ebd6a9ffc
+timeCreated: 1605643633
