--- Packages/UGF.Module.Update/Runtime/UpdateSystemDescriptionAssetBase.cs
@@ -1,8 +1,0 @@
-﻿using UGF.Builder.Runtime;
-
-namespace UGF.Module.Update.Runtime
-{
-    public abstract class UpdateSystemDescriptionAssetBase : BuilderAsset<IUpdateSystemDescription>
-    {
-    }
-}
