--- Packages/UGF.Module.Update/Runtime/UpdateSystemDescriptionAssetBase.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 10d61bf07f3d4bbea735055ebd6a9ffc
-timeCreated: 1605643633
