+++ Packages/UGF.Module.Update/Runtime/UpdateSystemDescriptionDefaultAsset.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ea71b94d072e454e90f7c508a0972ae1
+timeCreated: 1605643696
